1. 修正从服务器获取商品信息时格式错误无法反序列号json字符串的问题；2. 修改Log文件的存储路径到程序目录，修改Log记录格式。

+++ MarketSystem/ShopItem.cs
@@ -95,10 +95,10 @@ namespace MarketSystem
         /// <summary>
         /// 商品原价
         /// </summary>
-        public int itemOriginalPrice { get; set; }
+        public double itemOriginalPrice { get; set; }
         /// <summary>
         /// 商品当前售价
         /// </summary>
-        public int itemSellPrice { get; set; }
+        public double itemSellPrice { get; set; }
     }
 }
